+++ napari/components/viewer_model.py
@@ -810,6 +810,8 @@ class ViewerModel(KeymapProvider, MousemapProvider, EventedModel):
 
             return layer
         else:
+            from ..utils.stack_utils import split_channels
+
             warnings.warn(
                 trans._(
                     "\nThe 'channel_axis' argument is deprecated and will be removed in v0.5.1.\nTo achieve the same behavior, you may use 'napari.utils.split_channels':\n\n```\nfrom napari.utils import split_channels\n\nfor (ch, kw, _) in split_channels(data, channel_axis, **kwargs):\n    viewer.add_image(ch, **kw)\n```\n",
